hook up new pianoroll behavior
