+++ src/SelectSongPage/Modal.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { useRef, useState, useEffect } from 'react'
 import { PlayableSong, TrackSetting, TrackSettings } from '../types'
 import { useSelectedSong, cachedSettings } from '../hooks'
-import { SongVisualizer } from '../PlaySongPage'
+import { PianoRoll, SongVisualizer } from '../PlaySongPage'
 import { getHandSettings, applySettings, getTrackSettings } from '../PlaySongPage/utils'
 import { SongScrubBar } from '../pages/play/[...song_location]'
 import { getSong, inferHands, Sizer, formatInstrumentName } from '../utils'
