+++ src/utils.tsx
@@ -254,9 +254,19 @@ function diffKeys<T>(o1: T, o2: T): Array<keyof T> {
   return diff
 }
 
+function getNoteSizes(width: number, whiteCount: number) {
+  const whiteWidth = width / whiteCount
+  const whiteHeight = (5) * whiteWidth
+  const blackWidth = whiteWidth / 2
+  const blackHeight = whiteHeight * (2 / 3)
+
+  return { whiteWidth, whiteHeight, blackWidth, blackHeight }
+}
+
 export {
   Sizer,
   getSong,
+  getNoteSizes,
   formatTime,
   inferHands,
   Deferred,
