+++ tsconfig.json
@@ -6,7 +6,12 @@
     "jsx": "preserve",
     "sourceMap": true,
     "target": "es5",
-    "lib": ["dom", "dom.iterable", "esnext", "es2019"],
+    "lib": [
+      "dom",
+      "dom.iterable",
+      "esnext",
+      "es2019"
+    ],
     "allowJs": true,
     "skipLibCheck": true,
     "esModuleInterop": true,
@@ -20,6 +25,10 @@
     "downlevelIteration": true,
     "baseUrl": "."
   },
-  "include": ["src"],
-  "exclude": ["src\\PlaySongPage\\WindowedSongboard.tsx"]
+  "include": [
+    "src"
+  ],
+  "exclude": [
+    "node_modules"
+  ]
 }
